requirements.txt
fyers-apiv3
pandas
gcsfs
google-cloud-storage
python-telegram-bot
requests
pyarrow